# docs/swagger.yaml
basePath: /
definitions:
  http.LoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  http.LoginResponse:
    properties:
      access_token:
        type: string
      refresh_token:
        type: string
      user:
        properties:
          email:
            type: string
          id:
            type: integer
        type: object
    type: object
  http.RefreshRequest:
    properties:
      refresh_token:
        type: string
    required:
    - refresh_token
    type: object
  http.RefreshResponse:
    properties:
      access_token:
        type: string
      refresh_token:
        type: string
    type: object
  model.Ejercicio:
    properties:
      grupo_muscular_id:
        type: integer
      id:
        type: integer
      nombre:
        type: string
      tipo_ejercicio_id:
        type: integer
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: API para gestión de rutinas de gimnasio
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: GymBro API
  version: "1.0"
paths:
  /auth/login:
    post:
      consumes:
      - application/json
      description: Autentica un usuario y genera access token y refresh token
      parameters:
      - description: Credenciales de login
        in: body
        name: credentials
        required: true
        schema:
          $ref: '#/definitions/http.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/http.LoginResponse'
        "400":
          description: Datos inválidos
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Credenciales inválidas
          schema:
            additionalProperties: true
            type: object
      summary: Iniciar sesión
      tags:
      - autenticación
  /auth/logout:
    post:
      consumes:
      - application/json
      description: Invalida el refresh token (en producción, agregar a blacklist)
      parameters:
      - description: Refresh token a invalidar
        in: body
        name: refresh
        required: true
        schema:
          $ref: '#/definitions/http.RefreshRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Sesión cerrada exitosamente
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Datos inválidos
          schema:
            additionalProperties: true
            type: object
      summary: Cerrar sesión
      tags:
      - autenticación
  /auth/refresh:
    post:
      consumes:
      - application/json
      description: Renueva el access token usando un refresh token válido
      parameters:
      - description: Refresh token
        in: body
        name: refresh
        required: true
        schema:
          $ref: '#/definitions/http.RefreshRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/http.RefreshResponse'
        "400":
          description: Datos inválidos
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Refresh token inválido
          schema:
            additionalProperties: true
            type: object
      summary: Refrescar token
      tags:
      - autenticación
  /ejercicio:
    get:
      consumes:
      - application/json
      description: Obtiene la lista completa de ejercicios
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.Ejercicio'
            type: array
        "500":
          description: Error interno del servidor
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Obtener todos los ejercicios
      tags:
      - ejercicios
    post:
      consumes:
      - application/json
      description: Crea un nuevo ejercicio en el sistema
      parameters:
      - description: Datos del ejercicio
        in: body
        name: ejercicio
        required: true
        schema:
          $ref: '#/definitions/model.Ejercicio'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.Ejercicio'
        "400":
          description: Datos inválidos
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Error interno del servidor
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Crear nuevo ejercicio
      tags:
      - ejercicios
  /ejercicio/{id}:
    delete:
      consumes:
      - application/json
      description: Elimina un ejercicio del sistema
      parameters:
      - description: ID del ejercicio
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: Ejercicio eliminado
        "400":
          description: ID inválido
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Error interno del servidor
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Eliminar ejercicio
      tags:
      - ejercicios
    get:
      consumes:
      - application/json
      description: Obtiene un ejercicio específico por su ID
      parameters:
      - description: ID del ejercicio
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Ejercicio'
        "400":
          description: ID inválido
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Ejercicio no encontrado
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Obtener ejercicio por ID
      tags:
      - ejercicios
    put:
      consumes:
      - application/json
      description: Actualiza un ejercicio existente
      parameters:
      - description: ID del ejercicio
        in: path
        name: id
        required: true
        type: integer
      - description: Datos actualizados del ejercicio
        in: body
        name: ejercicio
        required: true
        schema:
          $ref: '#/definitions/model.Ejercicio'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Ejercicio'
        "400":
          description: Datos inválidos
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Error interno del servidor
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Actualizar ejercicio
      tags:
      - ejercicios
  /ejercicio/grupo-muscular/{id}:
    get:
      consumes:
      - application/json
      description: Obtiene todos los ejercicios de un grupo muscular específico
      parameters:
      - description: ID del grupo muscular
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.Ejercicio'
            type: array
        "400":
          description: ID inválido
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Error interno del servidor
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Obtener ejercicios por grupo muscular
      tags:
      - ejercicios
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"